+++ src/lib/check-auth.js
@@ -1,8 +1,6 @@
-const checkAuth = () => async (ctx, next) => {
+export default () => async (ctx, next) => {
   if (ctx.session && ctx.session.user) {
     return await next()
   }
   throw new Error('User not authorized')
 }
-
-module.exports = checkAuth