+++ src/types/Middleware.js
@@ -33,4 +33,4 @@
  */
 const m = {}
 
-module.exports = m
+export default m